fix(tcp): reading offset calculation
Since we start reading into the buffer at roff_, it needs to be
updated every time an incomplete packet is read, not just when
an entire packet was read. This bug causes the TCP connection to
reset every time an incomplete packet is read, since the TLV
parsing subsequently fails.

Signed-off-by: Varun Patil <varun.patil@operantnetworks.com>